requirements.txt
Flask
pandas
plotly
openpyxl
gunicorn
python-dotenv